ForeignEye_v2-main/backend/requirements.txt
# Flask 웹 프레임워크
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5

# 데이터베이스/ORM
PyMySQL==1.1.0
alembic==1.13.1

# 비동기 작업 & 캐시
celery==5.3.6
redis==5.0.1

# 그래프 DB (Neo4j)
neo4j==5.23.0

# HTTP 요청 및 웹 스크래핑
requests==2.31.0
httpx==0.27.0
beautifulsoup4==4.12.2

# 환경 변수 관리
python-dotenv==1.0.0

# AI/LLM
openai==1.3.0

# 검증
email-validator==2.1.0

# 보안/암호화
cryptography==43.0.0

# 프로덕션 서버
gunicorn==21.2.0

# (선택적) 개발 도구
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.12.0
# flake8==6.1.0
